refactor: add native-tls module
